Wcf layer created
WcfServiceLib is the new service layer, and it seems like it has
connection to the database now.

Removed "virtual" from ICollections, and it seems like it has an impact.

The old service layer - WcfServiceLibrary - seems broken and removing
"virtual" is not enough.

Added a connectionstring to the service layer since it kept going on a
localDB that it created itself.